feat: Add TabsCustomAnimation component for displaying account information
The code changes in the `page.tsx` file add a new component called `TabsCustomAnimation` for displaying account information. This component utilizes the `Tabs` and `TabPanel` components from the `@material-tailwind/react` library to create a tabbed interface with language options. The `AccountInfo` component is rendered within each tab to display the account information. This feature enhances the user account page by providing a visually appealing and organized way to view and switch between different language versions of the account information.