refactor: Update user model to include user_email_display field
The code changes in the `prisma/schema.prisma` file add a new field called `user_email_display` to the `user` model. This field is used to store the display name of the user's email address. This refactor enhances the user model by providing a separate field for displaying the email address, which can be useful for privacy or customization purposes.